# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "ERC1155 events.",
        "title": "ERC1155-events",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Jonathan Rwabahizi",
            "url": "http://www.swagger.io/support",
            "email": "jonathanrwabahizi@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/transactions": {
            "get": {
                "description": "get transactions list",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "get transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Transaction"
                            }
                        }
                    },
                    "404": {
                        "description": "no transactions found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/transactions/{hash}": {
            "get": {
                "description": "get single transaction",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "get single transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "transaction hash",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    },
                    "404": {
                        "description": "no transaction with hash: {hash} found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Transaction": {
            "type": "object",
            "required": [
                "blockHash",
                "hash"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "blockHash": {
                    "type": "string"
                },
                "blockNumber": {
                    "type": "string"
                },
                "contractAddress": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "gas": {
                    "type": "number"
                },
                "gasPrice": {
                    "type": "number"
                },
                "hash": {
                    "type": "string"
                },
                "receiverBalance": {
                    "type": "number"
                },
                "senderBalance": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
    }
}